Break out hour cycle into its own module
